fix : 409 status code when add_bookmark now returns a boorkmark object
I noticed that whether the documentation said that when a 409 status "should also provide a Location header to the resource in question", it was not the case when applying add_bookmark on an already processed url.
add_bookmark was raising an Error whenever the request status was not 200 or 202, instead of returning
the existing bookmark of corresponding to the requested Url; or at least it's location header.
Fixed it by adding the '409' status to the accepted add_bookmark requests status. Now returns
the already existing bookmark.